fix(trade-bench): fix conflict with main
